+++ src/Produtos/ProdutoService.java
@@ -10,7 +10,7 @@ public class ProdutoService {
     }
 
     // Vender produto pelo nome
-    public boolean vender(String nomeProduto, long quantidadeCliente) {
+    public boolean vender(String nomeProduto, float quantidadeCliente) {
         if (nomeProduto == null || nomeProduto.trim().isEmpty()) {
             System.out.println("Nome do produto inválido.");
             return false;
@@ -22,7 +22,7 @@ public class ProdutoService {
         }
 
         float precoUnitario = dao.obterPrecoPorNome(nomeProduto);
-        long estoqueAtual = dao.obterQuantidadePorNome(nomeProduto);
+        float estoqueAtual = dao.obterQuantidadePorNome(nomeProduto);
 
         if (precoUnitario == -1 || estoqueAtual == -1) {
             System.out.println("Produto não encontrado: " + nomeProduto);
@@ -34,7 +34,7 @@ public class ProdutoService {
             return false;
         }
 
-        long novoEstoque = estoqueAtual - quantidadeCliente;
+        float novoEstoque = estoqueAtual - quantidadeCliente;
         dao.atualizarQuantidadePosVenda(nomeProduto, novoEstoque);
 
         System.out.printf("Venda realizada! Produto: %s | Total a pagar: R$ %.2f | Novo estoque: %d%n",
@@ -51,7 +51,7 @@ public class ProdutoService {
         return dao.obterPrecoPorNome(nomeProduto);
     }
 
-    public void CadastrarProduto(String nome, float preco, String tipo, long quantidade) {
+    public void CadastrarProduto(String nome, float preco, String tipo, float quantidade) {
         if (nome == null || nome.trim().isEmpty()) return;
         if (preco <= 0 || quantidade < 0) return;
         if (dao.produtoExiste(nome.toUpperCase())) return;
@@ -64,7 +64,7 @@ public class ProdutoService {
         dao.inserir(p);
     }
 
-    public void EditarProduto(int id, String nome,float preco, String tipo, long quantidade) {
+    public void EditarProduto(int id, String nome,float preco, String tipo, float quantidade) {
         if (nome == null || nome.trim().isEmpty()) return;
         if (preco <= 0 || quantidade < 0) return;
